+++ tests/out-test.sh
@@ -1,0 +1,8 @@
+#!/bin/sh
+
+index=1
+while [ $index -le 1 ]
+do
+  echo "$* Data $index value1value1value1value1value1"
+  index=$(( $index + 1 ))
+done